+++ parser/CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Parse doc strings with `///` (or `//!` for inline structs and enums).
 - Support attributes for inline structs and enums (`#![...]`).
 - Added `Formatter` to format schema files.
+- Add `LinkResolver` for resolving doc links to nodes in the ast.
 
 ### Changed